Refactoring the classes into their respective packages
