+++ src/main/java/duke/launcher/Duke.java
@@ -1,11 +1,12 @@
-package duke;
+package duke.launcher;
 
 import duke.parser.ParserCommand;
 import duke.data.Storage;
-import duke.module.Schedule;
-import duke.sports.ManageStudents;
-import duke.sports.MyPlan;
+import duke.models.Schedule;
+import duke.models.ManageStudents;
+import duke.models.MyPlan;
 import duke.task.TaskList;
+import duke.view.CliView;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
