+++ src/main/java/duke/launcher/gui/package-info.java
@@ -2,4 +2,4 @@
  * Package holds all necessary code to initialize
  * the graphical user interface (gui).
  */
-package duke.gui;
+package duke.launcher.gui;
