+++ src/main/java/duke/models/Lesson.java
@@ -1,4 +1,4 @@
-package duke.module;
+package duke.models;
 
 import duke.data.Storage;
