+++ src/main/java/duke/models/ManageStudents.java
@@ -1,4 +1,4 @@
-package duke.sports;
+package duke.models;
 
 
 import java.util.ArrayList;
