+++ src/main/java/duke/models/Schedule.java
@@ -1,4 +1,4 @@
-package duke.module;
+package duke.models;
 
 import duke.data.Storage;
