+++ src/main/java/duke/models/package-info.java
@@ -8,4 +8,4 @@
  * 4. Training Programs
  */
 
-package duke.module;
+package duke.models;
