--- src/main/java/duke/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Package contains main program.
- */
-
-package duke;
