--- src/main/java/duke/sports/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Package contains all the basic building classes for
- * the class, plan, student, and training.
- */
-
-package duke.sports;
