+++ src/main/resources/view/menu.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.text.*?>
 
 <VBox prefHeight="720.0" prefWidth="1280.0" style="-fx-background-color: #43436e;" xmlns="http://javafx.com/javafx/8"
-      xmlns:fx="http://javafx.com/fxml/1" fx:controller="duke.gui.Controller">
+      xmlns:fx="http://javafx.com/fxml/1" fx:controller="duke.launcher.gui.Controller">
     <children>
         <MenuBar VBox.vgrow="NEVER">
             <menus>
