+++ src/main/resources/view/schedule.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.text.*?>
 
 <VBox prefHeight="720.0" prefWidth="1280.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-      fx:controller="duke.gui.Controller">
+      fx:controller="duke.launcher.gui.Controller">
     <children>
         <MenuBar prefHeight="26.0" prefWidth="600.0">
             <menus>
